requirements.txt
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
WTForms-Alchemy==0.16.9
gunicorn==20.0.4
itsdangerous==1.1.0
pytz==2019.3
psycopg2==2.8.4
python-dotenv==0.12.0
structlog==20.1.0
termcolor==1.1.0
werkzeug==1.0.0
WTForms==2.2.1
